pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lib-sensor-encoding"
version = "0.0.2"
dependencies = [
  "construct",
  "hexdump",
  "paho-mqtt",
]

[tool.setuptools.packages]
find = {}
